LLM/requirements.txt
google-cloud-bigquery
google-cloud-aiplatform
pandas
db-dtypes